moving checkbox
